# src/integration-test/resources/openapi.yml
openapi: 3.1.0
info:
  title: api-supportcenter
  contact: {}
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: "2.4"
servers:
  - url: http://localhost:63063
    description: Generated server url
...
paths:
...
  /{municipalityId}/endOfLeaseBatches:
    post:
      tags:
        - End of lease
      summary: Register a batch of computers that have reached end of lease
      operationId: createEndOfLeaseBatch
      parameters:
        - name: municipalityId
          in: path
          description: Municipality Id of the sender. The municipality of each computer
            is resolved from POB and may differ
          required: true
          schema:
            type: string
          example: 2281
        - name: adv-dataset-request
          in: header
          description: Set by the sender. Returned unchanged on every response
          schema:
            type: string
          example: DSET0001234
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEndOfLeaseBatchRequest"
        required: true
      responses:
        "202":
          description: Batch accepted for processing
          headers:
            Location:
              style: simple
              schema:
                type: string
            adv-dataset-request:
              description: "The adv-dataset-request header of the request, unchanged.\
                \ Left out when that header is missing or empty"
              style: simple
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EndOfLeaseBatchResponse"
        "409":
          description: "Conflict, a batch with the same external id is already registered.\
            \ The detail names the stored batch"
          headers:
            adv-dataset-request:
              description: "The adv-dataset-request header of the request, unchanged.\
                \ Left out when that header is missing or empty"
              style: simple
              schema:
                type: string
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad Request
          headers:
            adv-dataset-request:
              description: "The adv-dataset-request header of the request, unchanged.\
                \ Left out when that header is missing or empty"
              style: simple
              schema:
                type: string
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Problem"
                  - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server Error
          headers:
            adv-dataset-request:
              description: "The adv-dataset-request header of the request, unchanged.\
                \ Left out when that header is missing or empty"
              style: simple
              schema:
                type: string
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
  /{municipalityId}/endOfLeaseStatistics:
    get:
      tags:
        - End of lease
      summary: How the end of lease batches of a municipality have gone
      description: Answers with what the batches of a window add up to and with the same counts one batch at a time, so that a batch worth looking closer at can be picked out and read by its id. The window is the month up to today when none is given, and the one that was answered is part of the answer. The counts are the state the computers are in now. A computer that failed and then went through counts as sent
      operationId: getEndOfLeaseStatistics
      parameters:
        - name: municipalityId
          in: path
          description: Municipality Id of the sender that registered the batches
          required: true
          schema:
            type: string
          example: 2281
        - name: from
          in: query
          description: First day to count, by the day the batch was registered. A month back from the last day when left out
          required: false
          schema:
            type: string
            format: date
          example: 2026-08-18
        - name: to
          in: query
          description: Last day to count, counted in full. Today when left out
          required: false
          schema:
            type: string
            format: date
          example: 2026-09-18
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndOfLeaseStatisticsResponse'
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/Problem'
                  - $ref: '#/components/schemas/ConstraintViolationProblem'
        '500':
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /{municipalityId}/endOfLeaseBatches/{batchId}:
    get:
      tags:
        - End of lease
      summary: Read one batch with the computers in it
      description: Answers with the state each computer of the batch is in and, for the ones that were given up on, why. The reason is what the last attempt said, and a run that gets through clears it, so a computer that is SENT carries none even when it took several tries. The counts cover the whole batch whichever states are asked for
      operationId: getEndOfLeaseBatch
      parameters:
        - name: municipalityId
          in: path
          description: Municipality Id of the sender that registered the batch
          required: true
          schema:
            type: string
          example: 2281
        - name: batchId
          in: path
          description: Id of the batch, as answered by the call that registered it
          required: true
          schema:
            type: string
          example: 8f3c1e0a-2b4d-4f2e-9c7a-1d5e6f7a8b9c
        - name: status
          in: query
          description: States to list, one or more of PENDING, SENT, FAILED and EXCLUDED. Every state is listed when the parameter is left out
          required: false
          schema:
            type: array
            items:
              type: string
          example: FAILED
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndOfLeaseBatchStatusResponse'
        '400':
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/Problem'
                  - $ref: '#/components/schemas/ConstraintViolationProblem'
        '404':
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '500':
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
...
components:
  schemas:
...
    Problem:
      type: object
      properties:
        title:
          type: string
        detail:
          type: string
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        status:
          type: integer
          format: int32
    ConstraintViolationProblem:
      type: object
      properties:
        type:
          type: string
          format: uri
        status:
          type: integer
          format: int32
        violations:
          type: array
          items:
            $ref: "#/components/schemas/Violation"
        title:
          type: string
        detail:
          type: string
        causeAsProblem:
          $ref: "#/components/schemas/ThrowableProblem"
        instance:
          type: string
          format: uri
    ThrowableProblem:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        causeAsProblem: {}
    Violation:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
...
    CreateEndOfLeaseBatchRequest:
      type: object
      description: CreateEndOfLeaseBatchRequest model
      properties:
        externalBatchId:
          type: string
          description: "Id of the batch, set by the sender. A batch sent again under\
            \ an id that is already registered is rejected with 409 Conflict and not\
            \ queued a second time"
          examples:
            - DSET0001234
          maxLength: 36
          minLength: 1
        computers:
          type: array
          items:
            $ref: "#/components/schemas/EndOfLeaseComputer"
            description: Computers that have reached end of lease
          minItems: 1
      required:
        - computers
        - externalBatchId
    EndOfLeaseComputer:
      type: object
      description: EndOfLeaseComputer model
      properties:
        serialNumber:
          type: string
          description: Serial number
          examples:
            - J123ABC
          maxLength: 64
          minLength: 1
        assetTag:
          type: string
          description: Asset tag
          examples:
            - AB12345
          maxLength: 64
          minLength: 1
        endOfLeaseDate:
          type: string
          format: date
          description: "Date when the lease ends. Stored for traceability, does not\
            \ control when the message is sent"
          examples:
            - 2026-11-30
      required:
        - assetTag
        - endOfLeaseDate
        - serialNumber
    EndOfLeaseBatchResponse:
      type: object
      description: EndOfLeaseBatchResponse model
      properties:
        id:
          type: string
          description: Id of the registered batch
          examples:
            - 8f3c1e0a-2b4d-4f2e-9c7a-1d5e6f7a8b9c
        externalBatchId:
          type: string
          description: "Id of the batch, set by the sender"
          examples:
            - DSET0001234
    EndOfLeaseStatisticsResponse:
      type: object
      description: EndOfLeaseStatisticsResponse model
      properties:
        from:
          type: string
          format: date
          description: "First day the counts cover. The window that was asked for,\
            \ or the month up to today when none was"
          examples:
            - 2026-08-18
        to:
          type: string
          format: date
          description: "Last day the counts cover, counted in full"
          examples:
            - 2026-09-18
        batches:
          type: integer
          format: int32
          description: Number of batches the municipality registered inside the window
          examples:
            - 128
        computers:
          $ref: "#/components/schemas/EndOfLeaseComputerCounts"
          description: "The computers of every batch, counted per state"
        perBatch:
          type: array
          items:
            $ref: "#/components/schemas/EndOfLeaseBatchStatistics"
            description: "The same counts for one batch at a time, newest batch first"
    EndOfLeaseComputerCounts:
      type: object
      description: EndOfLeaseComputerCounts model
      properties:
        total:
          type: integer
          format: int64
          description: "Number of computers that have been received, which is the\
            \ sum of the states below"
          examples:
            - 1240
        pending:
          type: integer
          format: int64
          description: "Number of computers waiting to be sent, or waiting to be sent\
            \ again after an attempt that did not succeed"
          examples:
            - 12
        sent:
          type: integer
          format: int64
          description: Number of computers that have been sent their message
          examples:
            - 1180
        failed:
          type: integer
          format: int64
          description: "Number of computers whose attempts are used up, which need\
            \ to be looked at by a human"
          examples:
            - 8
        excluded:
          type: integer
          format: int64
          description: "Number of computers that are never sent a message, recognized\
            \ by the letters their asset tag starts with"
          examples:
            - 40
    EndOfLeaseBatchStatistics:
      type: object
      description: EndOfLeaseBatchStatistics model
      properties:
        id:
          type: string
          description: "Id of the batch, which is what a single batch is read by"
          examples:
            - 8f3c1e0a-2b4d-4f2e-9c7a-1d5e6f7a8b9c
        externalBatchId:
          type: string
          description: Id the sender gave the batch
          examples:
            - d1f3a8c2-9b7e-4a5f-8c3d-2e6b1a4f7c90
        created:
          type: string
          format: date-time
          description: When the batch was registered
          examples:
            - 2026-09-17T06:03:11+02:00
        total:
          type: integer
          format: int64
          description: Number of computers the batch was received with
          examples:
            - 982
        pending:
          type: integer
          format: int64
          description: Number of computers in the batch waiting to be sent
          examples:
            - 0
        sent:
          type: integer
          format: int64
          description: Number of computers in the batch that have been sent their
            message
          examples:
            - 961
        failed:
          type: integer
          format: int64
          description: Number of computers in the batch whose attempts are used up
          examples:
            - 3
        excluded:
          type: integer
          format: int64
          description: Number of computers in the batch that are never sent a message
          examples:
            - 18
    EndOfLeaseBatchStatusResponse:
      type: object
      description: EndOfLeaseBatchStatusResponse model
      properties:
        id:
          type: string
          description: Id of the batch
          examples:
            - 8f3c1e0a-2b4d-4f2e-9c7a-1d5e6f7a8b9c
        externalBatchId:
          type: string
          description: Id the sender gave the batch
          examples:
            - d1f3a8c2-9b7e-4a5f-8c3d-2e6b1a4f7c90
        created:
          type: string
          format: date-time
          description: When the batch was registered
          examples:
            - 2026-09-17T06:03:11+02:00
        total:
          type: integer
          format: int64
          description: Number of computers the batch was received with
          examples:
            - 982
        pending:
          type: integer
          format: int64
          description: Number of computers in the batch waiting to be sent
          examples:
            - 0
        sent:
          type: integer
          format: int64
          description: Number of computers in the batch that have been sent their
            message
          examples:
            - 961
        failed:
          type: integer
          format: int64
          description: Number of computers in the batch whose attempts are used up
          examples:
            - 3
        excluded:
          type: integer
          format: int64
          description: Number of computers in the batch that are never sent a message
          examples:
            - 18
        computers:
          type: array
          items:
            $ref: "#/components/schemas/EndOfLeaseComputerStatus"
            description: "Every computer in the batch, in state order and then by\
              \ serial number"
    EndOfLeaseComputerStatus:
      type: object
      description: EndOfLeaseComputerStatus model
      properties:
        serialNumber:
          type: string
          description: Serial number
          examples:
            - J123ABC
        assetTag:
          type: string
          description: "Asset tag, which is the name the computer is known by in SysMan"
          examples:
            - AB12345
        status:
          type: string
          description: "The state the computer is in: PENDING, SENT, FAILED or EXCLUDED.\
            \ A string rather than an enumeration, so that a state added later reaches\
            \ a client that has not heard of it yet as a value it does not recognize\
            \ rather than as a broken answer"
          examples:
            - FAILED
        attempts:
          type: integer
          format: int32
          description: "How many attempts the computer has cost. An attempt is only\
            \ spent on something that is the computer's own fault, so a dependency\
            \ that could not be reached does not show here"
          examples:
            - 5
        errorMessage:
          type: string
          description: "Why the last attempt did not succeed. Cleared as soon as an\
            \ attempt gets through, so a computer that is SENT carries none even when\
            \ it took several tries to get there"
          examples:
            - POB answered the serial number with no configuration item holding a municipality
              we can route on
        sentAt:
          type: string
          format: date-time
          description: "When the message was sent, or null when it has not been"
          examples:
            - 2026-09-17T06:14:52+02:00
...
  securitySchemes: {}